[multi-cluster] Resolve clusters available to the adjacent istio control plane (#3413)
* Function to discover what remote clusters are visible and also discover the "home" cluster

"Home cluster" means: the cluster where the "adjacent" control plane is running. "Adjacent control plane" means the control plane that Kiali is observing/managing.

Remote clusters are discovered by fetching the list of "remote secrets" that are present in the istio namespace.

Apparently, remote secrets are labeled with istio/multiCluster=true (at least, this is the label that presets the `istioctl x create-remote-secret` command). So, this is fetching secrets with that label and assuming they contain the kubeconfig files required by istio to contact the remote API Server and enable discovery.

Related kiali/kiali#3409